Create model for players
Each player has a vehicle. Therefore, the Vehicle model is associated with the Player model. Player has-a Vehicle. The vehicle model is in a separate repository. If I recall correctly, this is where Graphene Federation comes into play. Or, perhaps, Node.js is to be utilized here. My memory is a bit hazy. Thus, for the sake of efficiency, the priority now is to simply get Player to display in the administration end of Django. Vehicle model: https://github.com/JoshuaJonesTS/racing-game-vehicles/blob/main/vehiclesapp/models.py