Add vehicle field to Player.
I'm going to assume that models.PositiveSmallIntegerField() is to be utilized for all relations for Tailored Software. It seems as though the data type of an id is not relevant to the data type that is used to create the association, as demonstrated by the microservice example. https://github.com/tailored-software/djangoexample1/blob/staging/customers/migrations/0001_initial.py. This model was used as the template: https://github.com/tailored-software/djangoexample2/blob/staging/appts/models.py,